requirements.txt
h5py==2.8.0
joblib==0.9.4
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
libtiff==0.4.2
matplotlib==2.2.3
numpy==1.14.5
opencv-python==3.4.3.18
openslide-python==1.1.1
pandas==0.17.1
pip==19.0.3
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.2.1
setuptools==39.1.0
tensorboard==1.10.0
tensorflow==1.10.1
tifffile==2019.3.18
tqdm==4.28.1
